Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21816 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/impexp/ase/aseie.cpp
@@ -405,7 +405,7 @@ bool csASEInterpreter_MESH_NORMALS (csModelConverterASE *conv,
   if (CS_ASE_CHECK_TOKEN ("*MESH_FACENORMAL"))
   {
     int const n = in.ReadTextInt ();
-    if (n < 0 || n >= conv->Polygons.Length ()) return false;
+    if (n < 0 || (size_t)n >= conv->Polygons.Length ()) return false;
     conv->CurrentPolygon = conv->Polygons.Get ((size_t)n);
     conv->CurrentVertex = 0;
     return true;
